fix: tests and lint errors
