+++ .eslintrc.json
@@ -46,6 +46,8 @@
     "import/prefer-default-export": "off",
     "@typescript-eslint/no-explicit-any": "error",
     "@typescript-eslint/ban-ts-comment": "off",
+    "no-useless-constructor": "off",
+    "no-empty-function": "off",
     "node/no-missing-import": [
       "error",
       {
