+++ package.json
@@ -64,6 +64,8 @@
     "@nestjsx/crud-typeorm": "^5.0.0-alpha.3",
     "@types/pino-multi-stream": "^5.1.3",
     "bcrypt": "^5.0.1",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.13.2",
     "cross-env": "7.0.3",
     "dotenv": "10.0.0",
     "fastify": "^3.24.1",
@@ -93,6 +95,7 @@
     "@types/bcrypt": "^5.0.0",
     "@types/express": "^4.17.13",
     "@types/jsonwebtoken": "^8.5.8",
+    "@types/multer": "^1.4.7",
     "@types/node": "^16.11.12",
     "@types/passport": "^1.0.7",
     "@types/passport-jwt": "^3.0.6",
