+++ src/app.service.ts
@@ -1,3 +1,5 @@
+/* eslint-disable class-methods-use-this */
+
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
