+++ src/auth/auth.module.ts
@@ -7,7 +7,6 @@ import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 
 import ENV from '../common/config';
-import { UserService } from '../user/user.service';
 
 @Module({
   imports: [
