+++ src/board/dto/createBoard.dto.ts
@@ -6,23 +6,13 @@ import {
   ValidateNested,
 } from 'class-validator';
 import { Type } from 'class-transformer';
-import { EColumn } from './column.entity';
+import { EColumn } from '../column.entity';
 
 export class CreateBoardDto {
   @IsOptional()
   @IsUUID()
   id!: string;
 
-  @Length(5, 30)
-  title!: string;
-
-  @IsArray()
-  @ValidateNested({ each: true })
-  @Type(() => EColumn)
-  columns!: EColumn[];
-}
-
-export class UpdateBoardDto {
   @Length(5, 30)
   title!: string;
