+++ src/user/dto/createUser.dto.ts
@@ -1,0 +1,7 @@
+export class CreateUserDto {
+  id!: string;
+
+  name!: string;
+
+  login!: string;
+}
