+++ src/user/dto/updateUser.dto.ts
@@ -1,11 +1,3 @@
-export class CreateUserDto {
-  id!: string;
-
-  name!: string;
-
-  login!: string;
-}
-
 export class UpdateUserDto {
   name!: string;
